requirements.txt
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
google-auth==2.32.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
httplib2==0.22.0
idna==3.7
oauth2client==4.1.3
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pyTelegramBotAPI==4.21.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
telebot==0.0.5
urllib3==2.2.2
